portfolio-backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "litellm>=1.81.1",
    "pydantic[email]>=2.12.5",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
